Move OffsetNode to the correct namespace
