+++ SAGESharp.Tests/IO/Binary/TreeBasedSerialization/OffsetNodeTests.cs
@@ -9,9 +9,10 @@ using NSubstitute.ClearExtensions;
 using NUnit.Framework;
 using SAGESharp.IO;
 using SAGESharp.IO.Binary;
+using SAGESharp.IO.Binary.TreeBasedSerialization;
 using System;
 
-namespace SAGESharp.Tests.IO
+namespace SAGESharp.Tests.IO.Binary.TreeBasedSerialization
 {
     class OffsetNodeTests
     {
