+++ SAGESharp.Tests/SAGESharp.Tests.csproj
@@ -86,13 +86,13 @@
     <Compile Include="IO\Binary\BinaryWriterWrapperTests.cs" />
     <Compile Include="IO\Binary\ReaderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\ListNodeTests.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\OffsetNodeTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBinarySerializerTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBuilderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeReaderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeWriterTests.cs" />
     <Compile Include="IO\Binary\WriterTests.cs" />
     <Compile Include="IO\EdgeTests.cs" />
-    <Compile Include="IO\OffsetNodeTests.cs" />
     <Compile Include="IO\PaddingNodeTests.cs" />
     <Compile Include="IO\PrimitiveTypeDataNodeTests.cs" />
     <Compile Include="IO\StringDataNodeTests.cs" />
