+++ SAGESharp/SAGESharp.csproj
@@ -97,6 +97,7 @@
     <Compile Include="IO\Binary\Reader.cs" />
     <Compile Include="IO\Binary\RightPaddingAttribute.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\ListNode.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\OffsetNode.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBinarySerializer.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBuilder.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeReader.cs" />
